Move wrapError to telemetry-utils package
