+++ api-report/container-utils.api.md
@@ -81,12 +81,6 @@ export class UsageError extends LoggingError implements IFluidErrorBase {
     readonly fluidErrorCode: string;
 }
 
-// @public
-export function wrapError<T extends IFluidErrorBase>(innerError: unknown, newErrorFn: (message: string) => T): T;
-
-// @public
-export function wrapErrorAndLog<T extends IFluidErrorBase>(innerError: unknown, newErrorFn: (message: string) => T, logger: ITelemetryLogger): T;
-
 
 // (No @packageDocumentation comment for this package)
