+++ api-report/telemetry-utils.api.md
@@ -247,6 +247,12 @@ export class ThresholdCounter {
     sendIfMultiple(eventName: string, value: number): void;
     }
 
+// @public
+export function wrapError<T extends IFluidErrorBase>(innerError: unknown, newErrorFn: (message: string) => T): T;
+
+// @public
+export function wrapErrorAndLog<T extends IFluidErrorBase>(innerError: unknown, newErrorFn: (message: string) => T, logger: ITelemetryLogger): T;
+
 
 // (No @packageDocumentation comment for this package)
