+++ packages/loader/container-utils/src/error.ts
@@ -10,12 +10,12 @@ import {
     IThrottlingWarning,
 } from "@fluidframework/container-definitions";
 import {
-    extractLogSafeErrorProperties,
     LoggingError,
     isValidLegacyError,
     IFluidErrorBase,
     normalizeError,
-    hasErrorInstanceId,
+    wrapError,
+    wrapErrorAndLog,
 } from "@fluidframework/telemetry-utils";
 import { ITelemetryLogger, ITelemetryProperties } from "@fluidframework/common-definitions";
 import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions";
@@ -69,6 +69,7 @@ export class ThrottlingWarning extends LoggingError implements IThrottlingWarnin
     ): IThrottlingWarning {
         const newErrorFn =
             (errMsg: string) => new ThrottlingWarning(errMsg, errorCode, retryAfterSeconds);
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-return
         return wrapErrorAndLog(error, newErrorFn, logger);
     }
 }
@@ -137,6 +138,7 @@ export class DataProcessingError extends LoggingError implements IErrorBase, IFl
         if (message !== undefined) {
             error.addTelemetryProperties(extractSafePropertiesFromMessage(message));
         }
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-return
         return error;
     }
 }
@@ -154,57 +156,3 @@ export const extractSafePropertiesFromMessage = (message: ISequencedDocumentMess
  * Conditionally coerce the throwable input into a DataProcessingError.
  */
  export const CreateProcessingError = DataProcessingError.wrapIfUnrecognized;
-
-/**
- * Create a new error, wrapping and caused by the given unknown error.
- * Copies the inner error's message and stack over but otherwise uses newErrorFn to define the error.
- * The inner error's instance id will also be logged for telemetry analysis.
- * @param innerError - An error from untrusted/unknown origins
- * @param newErrorFn - callback that will create a new error given the original error's message
- * @returns A new error object "wrapping" the given error
- */
-export function wrapError<T extends IFluidErrorBase>(
-    innerError: unknown,
-    newErrorFn: (message: string) => T,
-): T {
-    const {
-        message,
-        stack,
-    } = extractLogSafeErrorProperties(innerError, false /* sanitizeStack */);
-
-    const newError = newErrorFn(message);
-
-    if (stack !== undefined) {
-        // supposedly setting stack on an Error can throw.
-        try {
-            Object.assign(newError, { stack });
-        } catch (errorSettingStack) {
-            newError.addTelemetryProperties({ stack2: stack });
-        }
-    }
-
-    if (hasErrorInstanceId(innerError)) {
-        newError.addTelemetryProperties({ innerErrorInstanceId: innerError.errorInstanceId });
-    }
-
-    return newError;
-}
-
-/** The same as wrapError, but also logs the innerError, including the wrapping error's instance id */
-export function wrapErrorAndLog<T extends IFluidErrorBase>(
-    innerError: unknown,
-    newErrorFn: (message: string) => T,
-    logger: ITelemetryLogger,
-) {
-    const newError = wrapError(innerError, newErrorFn);
-    const wrappedByErrorInstanceId = hasErrorInstanceId(newError)
-        ? newError.errorInstanceId
-        : undefined;
-
-    logger.sendTelemetryEvent({
-        eventName: "WrapError",
-        wrappedByErrorInstanceId,
-    }, innerError);
-
-    return newError;
-}
